requirements.txt
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
Pygments==2.19.2
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pytesseract==0.3.13
rich==14.2.0
textual==6.3.0
typing_extensions==4.15.0
uc-micro-py==1.0.3